backend/scene_discription_2/requirements.txt
Flask==2.3.2
torch==2.0.1
transformers==4.30.2
gTTS==2.3.2
Pillow==9.5.0
numpy==1.26.4
googletrans==4.0.0-rc1